+++ lib/proj/context.rb
@@ -275,24 +275,5 @@ module Proj
       @network_api = network_api_klass.new(self)
     end
 
-    # Returns the path to the Proj database.
-    #
-    # @deprecated Use {Database#path} instead via {#database}. Will be removed in a future version.
-    # @return [String] The database path
-    def database_path
-      self.database.path
-    end
-
-    # Sets the path to the Proj database.
-    #
-    # @deprecated Use {Database#path=} instead via {#database}. Will be removed in a future version.
-    # @param value [String] The new database path
-    def database_path=(value)
-      self.database.path = value
-    end
-
-    extend Gem::Deprecate
-    deprecate :database_path, "context.database.path", 2023, 6
-    deprecate :database_path=, "context.database.path=", 2023, 6
   end
 end